scheduler: Admin task check: Optimized snapctl arguments
Changed the snapctl call from

snapctl is-connected --apparmor-label=AA_CONTEXT CUPS_CONTROL_SLOT

to

snapctl is-connected --apparmor-label AA_CONTEXT CUPS_CONTROL_SLOT

removing the '=' between "--apparmor-label" and "AA_CONTEXT" to make
these two separate arguments and so avoid the need to concatenate the
strings in a buffer.

Did the same with the arguments for the
snapd_client_run_snapctl2_sync() function call.