PPD option preset auto-generation for common job IPP attributes
To make the job IPP attributes "print-color-mode", "print-quality",
and "print-content-optimize" actually doing "the right thing" for most
printers we auto-generate PPD option presets now.

CUPS already had the functionality to apply a preset of one or more
PPD option settings for each of the 6 combinations of
"print-color-mode" and "print-quality" settings for longer time but
this was never actually being made use of as it required manually
defining the presets via "APPrinterPreset" PPD attributes. This no one
wants to do with the ~10000 PPDs which come with a typical Linux
distribution nowadays. Even a user with his few print queues does not
want to do it, know how to do it, or know that it is even possible.

This commit makes this fully automatic on every PPD where the presets
are not already manually defined via "APPrinterPreset" attributes.

When creating the PPD cache for a queue's PPD file and the PPD is
without manual presets, it calls the new function
_ppdCacheAssignPresets() which analyses for each PPD option the names
of the choices to find out which choices switch to manochrome or to
color, lower or raise print quality, or improve text/graphics/photo
printing and which are the moset suitable of them. It also analyses
the PostScript/PJL code of the choices to find out whether it affects
the printing resolution.

With this it selects which options get into the the presets and with
which choices.

In addition to presets for the "print-color-mode"/"print-quality"
combos presets for "print-content-optimize" are introduced.

PPD Option settings added by preset are logged on job execution.