Do not open logfiles with read access
Currently logfiles are opened with read-access, due to cupsFileOpen()
converting "a" into `O_RDWR | O_CREAT | O_APPEND`. This should not be
necessary, as logfiles should be an append-only data structure.
Update cupsFileOpen() to open files with `O_WRONLY | O_CREAT | O_APPEND`
on mode "a" and add a fallback mode "a+" (see man:fopen(3)).
Currently only the logfiles and a test file are opened with "a".