Fix crash when DNSSDHostName is NULL
If cupsd is running without Browsing (when not sharing printers) the
global variable DNSSDHostName of the scheduler is not set, staying
NULL. Therefore the previous commit causes a crash with this
configuration of CUPS.

This commit does a NULL check on DNSSDHostName and falls back to
ServerName if needed, where it also does a NULL check and if this is
also NULL, it refrains from any attempt to correct the device URI's
hostname to "localhost".

As ServerName usually contains the server's hostname without ".local"
suffix but the device URIs coming from Avahi usually have a suffixed
host name (and there are also the ".local" and ".local." variants of
the suffix), we consider hostnames also as equal if one has the
suffix, the other not, or if we have the variants with and without
trailing dot.

In addition, there are also HAVE_DNSSD conditionals added around the
code using DNSSDHostName now.