config-scripts: Fix --enable-libtool-unsupported
When the LIBTOOL variable is set in the user's environment the build
will use libtool even when --enable-libtool-unsupported is not used.

This can lead to unexpected failures when LIBTOOL is set to rlibtool
rather than GNU libtool.

To solve this the build now relies on the enable_libtool_unsupported
varaiable set by --enable-libtool-unsupported.

Gentoo bug: https://bugs.gentoo.org/show_bug.cgi?id=843638