choose-device.tmpl: Add device uri into title of found network printer
There can be many printers on the same network with the same printer
model, which causes confusion for users about which one to add. This PR
adds a title for every found network printer, which contains the
printer's device uri. The title is shown when user hover the mouse over
the found entry.

Partially fixes #393, thanks to Pro-pra for the PR!