Test suite: If holding supported, add a new print job for cancelling
On some quicker machines the print-job operation finishes quicker than
we can cancel it in the test - so add job-hold-until job template
attribute if the printer supports it, which will hold the job until it
is cancelled by the next test.

The next test passes because it has both statuses (correct and error one)
as expected, but cupsd generates an additional error in log if the job
cancelling fails, which would be more dirty to cover in run-stp-test.sh.

There is a loophole when the error is still generated - if the printer/machine
is quick enough to finish the job, but the machine/printer doesn't support job
holding at the same time. IMHO the current patch should cover most cases.