scheduler/ipp.c: Convert incoming ColorModel attribute
If a client uses PPD option ColorModel (which is converted into
attribute to be sent via IPP) for changing the default color option
(which happens via CUPS-Add-Modify-Printer operation), we have to
convert ColorModel into print-color-mode attribute, so the
print-color-mode-default attribute gets updated.

The change survives reboot because print-color-mode is saved as an
option in printer's entry in /etc/cups/printers.conf.