Add support for filtering the list of printers on the client side
When a cups server has hundreds of printers defined, each client-side operation becomes very slow, as the full list of printers is queried from the server for almost every operation. Also, the list of printers presented to the user becomes far too long, e.g. in GUI-based programs such as LibreOffice.

With this patch, a user can list a set of printers that {s}he wants to queried. This list can be specified using either an environment variable
export CUPS_PRINTER_LIST="printer1, printer2, printer3"
or by adding a line to the ~/.cups/lpoptions file
printer-list printer1, printer2, printer3

Please note that the list of printers can either be comma-separated or space-separated.

Co-Authored-By: jjkeijser <7151077+jjkeijser@users.noreply.github.com>