Change cupsd.conf to force authentication when accessing administration
With the current default configuration, anyone can request to Find New Printers without being authenticated and with no way to be authenticated. Without authentication cups does not detects printers on the network. The person might not realize that it's because authentication is missing. It seems natural to always offer to authenticate when accessing the administration page.

Co-Authored-By: Dominic Mayers <dominic.mayers@meditationstudies.org>