Improve state handling in cases of memory allocation failure
Many lines of code assume that malloc will not fail. In cases where it does, sometimes the program does not know, and as a result, memory can leak and more disastrous consequences can happen before the program ultimately finds something is wrong and then calls exit();

This PR aims to better handle memory in these cases.