Update InputSlot selection for photo sizes (fixes #569)
_ppdCacheGetInputSlot specifically requests the photo tray for paper
sizes up to 5x7" if the client hasn't specified a media-source.  This
can fail if the photo tray is physically smaller than 5x7 (for example,
L-sized).

Instead of explicitly requesting the photo tray, try to request "auto"
and fall back to "photo" if the printer doesn't provide that as an
option.

Signed-off-by: Benjamin Gordon <bmgordon@chromium.org>