Add templates for issue reports
This adds a selection screen when you click the button to report a new
issue on the "Issues" page. The selection screen does not only create
one entry for each issue template but also one extra entry for
security (private) bug reports. The latter is especially important as
otherwise the entrance point for a security bug report is hidden on
the security page.