cups/http-addr.c: Set listen backlog size to INT_MAX (fixes #308)
Use a listen queue size of INT_MAX, which should default to the maximum
supported queue size on the system.

This avoids the problem of the listening backlog queue getting full when
there are too many requests at the same time. The problem was observed
with the previous backlog size (128) by customers when submitting large
batches of print jobs, resulting in some jobs getting lost.

Signed-off-by: Vasilis Liaskovitis <vliaskovitis@suse.com>