Create shims to pass to array structures as to not pass incompatible function pointers
Although it may seem to work on modern systems, casting functions that do not have the same number of arguments as the expected type can cause ABI problems and will be removed from the latest version of C.

By sheer luck, most modern ABIs and compilers will treat the CPU register that contains the third argumentas a scratch register that may contain information from the function that called itand push it to the stack just in case. However, not all ABIs are guaranteed to do this, and because this program is not supposed to rely on these implementation details, it is best to ensure by the C standard that undefined behavior, or platform-dependent behavior, does not occur by the creation of private shim functions meant to be passed.

The reason I did this instead of changing the function definition is to maintain backwards compatibility; some of the functions being passed are public API and therefore I cannot change them.

This is the best I could do, but if there is anything else I could do, please let me know.