cups/dest.c: Raise timeout _CUPS_DNSSD_GET_DESTS
The current timeout is not able to list all network devices if there are
many IPP services on mDNS (the tested number is 165 services).

Raising the timeout to 1s does not slow libcups if there are less
services (Avahi returns earlier) or if Avahi does not run on the system
(libcups cannot create an Avahi client in that case), and provides time
frame for getting reasonable amount of IPP services (big enterprise
servers will use permanent queues and printer profiles than mDNS).

Fixes #751