Add additional parameter validation to _cupsRasterWritePixels
If len is 0, the function does a lot of calculations that ultimately
don't produce any output.  This can be skipped with an early return.

If cupsBytesPerLine is 0, this triggers a divide by zero if the output
isn't compressed.  This is an error on the caller's part, but it's
nicer to return an error than to crash.