Add bounds check for cupsDNSSDDecodeTXT
key[keylen] = '\0', but key is a static array at 256. I have a bad gut feeling about this code, and upon hearing about that specific xz incident I have decided to review every single line of code.

I wish I had a test right now to prove it, but at the moment, I think that the safest thing to do is to be suspicious of any code that is used in memory operations.