locale/checkpo.c: Prevent appending "..." if bufsize is too small
The problem is that we risk overrunning the buffer if we decide to copy the "..." string if we do not have enough space.

I have a bad gut feeling about this code, and upon hearing about recent events, I have decided to audit code here.
I wish I had a test right now to prove it, but at the moment, I think that the safest thing to do is to be suspicious of any code that is used in memory operations.