cgi-bin/var.c: Fix return value if the text is invalid
If the original value contains double quote, which is forbidden for
text, we free the variable and move the next variable from array to its
place. However, the return value still contains the original value of
freed pointer.

Set it NULL, since we don't have any value to return for the requested
name.