Work aroud broken {text,name}WithLanguage attributes
Some broken IPP implementations (noticed on a HP LaserJet 2430dtn), in a
textWithLanguage/nameWithLanguage attribute, omit the total attribute length.
Try to identify the broken case by looking at the first two bytes after what
should be the total length: normally, these are language-length, but in the
broken case are the first two characters of language instead.
In that case, compute the total length from language-length and
{text,name}-length.