cups/dest-options.c: Fix potential overwriting of pwg in cups_get_media_db
When a non-standard media name or size is used in cupsGetDestMediaByName
or cupsGetDestMediaBySize, the first call might return an incorrect size. This is
due to a potential overwriting of the pwg variable. This commit fixes the issue
by ensuring the pwg value is re-fetched after cups_create_media_db or
cups_update_ready is called.

Fixes #993