Create new function cgiGetVariablePtr that avoids allocation
The problem, cgiGetVariable calls strdup on success, which leaks memory all over the place.

This is going to be a lot of work to fix, but based on how CoreFoundation fixed this issue, I think just returning a pointer should work.

Undefined behavior happens if the user uses the returned pointer to write.