Support bare staple and bind finishings values
In the generated PPD, IPP finishings "bind" is supposed to map to
"StapleLocation: BindAuto".  Similarly, "staple" is supposed to map to
"StapleLocation: SingleAuto".  The code already handles this, except
the lookup is blocked by a check that only accepts "staple-*" and
"bind-*" prefixed versions.  Fix this by adding the bare versions to the
existing checks.

Fixes issue #1073.