libcups: Don't overwrite permanent queue with network during discovery
Currently, if there is a permanent queue of the same name as network
discovered printer, the network one is written in `lpstat -l -e` output.
This is confusing to users and maintainers, because if print job is
sent, it comes to the permanent one instead of listed network/temporary,
which can have a different connection or PPD.

This behavior happens when user creates manually a queue with the same
name as the temporary one, or when the permanent queue is created by
cups-browsed. In both cases user realizes he prints into a different
queue once he enables debugging and see printer-uri in logs.