scheduler: Fix applying print-as-raster default
Currently, the default was applied only when present in the request,
which is how `lp` behaves. However the application does not have to
include defaults in the request, so we have to apply it in scheduler.

Additionally, this covers a case when print-as-raster is set by default,
but we want to disable it in a specific request.