ppd-mark: try to map PageSize to PPD if media is missing. Fixes #1375
When CUPS prints to another CUPS over IPP, it uses PageSize instead of
media or media-col, as it is assumed that PPD files (or at least
the attribute values) are synced on both sides.

However this is not the case with ippeve or drvless auto-generated
PPDs, where only Adobe names are used for page sizes instead of the
values on the remote side, which makes page selection inconsistent.

Try to map Adobe names to PPD names on the server side, if only
PageSize is provided in IPP request.