Fix src shift in ipp-file.c
Each case (`$$`, `$ENV[`, `${`, and plain `$`) is supposed to shift `src` after parsing. But the block responsible for the `$` case lacks this shift--it is located out of that block. So for the `${` case `src` is shifted twice: by `src += 2` and by `src += tempptr - temp + 1`, which results in `while (*src)` reading beyond the NUL-terminator.